md_to_html.sh: Make premailer optional
It turns out that premailer is changing some of the output styling, and
since it is not currently being copy-pasted into Webcourses anymore,
premailer will not be used unless specified. Premailer will still be
kept in the script for those cases where you do want to paste HTML into
Webcourses.

Signed-off-by: Jerrett Longworth <jerrett7@gmail.com>