README.md: Consolidate the table of contents
Since there are now really two versions of the table of contents,
the one in the README and the one in the index.html files, adding new
materials becomes a somewhat redundant task. This consolidates the
table of contents into their respective subjects, with only the main
topic categories being linked.

The previous two commits should allow the casual GitHub browser to
click on the links in the new locations properly.