Update numbered lists to generic numbering
I have finally discovered how to create lists that don't have to be
manually renumbered anymore. No longer does a list have to be renumbered
each time a new problem is added to a list of practice problems. Pandoc
has this feature in their user guide:
https://pandoc.org/MANUAL.html#numbered-example-lists

The wiki will also be updated to reflect this.

Signed-off-by: Jerrett Longworth <jerrett7@gmail.com>