Update switch-statements-practice-answers.md
This should resolve the very cryptic issue that I posted. The answer has been cleaned of its whitespace, and I added a note about the outputs. Seeing as the inputs given are separated by whitespace to emphasize they are different runs, I noted that each output should appear on a new line. Please look through and check my answer key. If it's good, it should be merge ready.